Separated null_strand as a file.
